# openapi.yaml
openapi: 3.0.0
info:
  title: RealChat API
  version: 1.0.0
paths:
  /chat:
    get:
      summary: Retrieve messages from a given chat ID
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The chat ID to retrieve messages for
      responses:
        '200':
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '400':
          description: Bad request when the ID is not valid
        '500':
          description: Internal server error when the chat cannot be retrieved
  /new:
    get:
      summary: Create a new chat
      responses:
        '200':
          description: The ID of the newly created chat
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error when the chat cannot be created
  /send:
    post:
      summary: Send a message to a chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomingMessage'
      responses:
        '200':
          description: The ID of the chat to which the message was sent
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request when the message data is not valid
        '500':
          description: Internal server error when the message cannot be sent
components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint32
        text:
          type: string
        timestamp:
          type: string
          format: date-time
        chat_id:
          type: integer
          format: uint32
        sender_id:
          type: integer
          format: uint32
    IncomingMessage:
      type: object
      properties:
        chat_id:
          type: integer
          format: uint32
        sender_id:
          type: integer
          format: uint32
        text:
          type: string